requirements.txt
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
requests==2.31.0
jsonschema==4.20.0
python-dotenv
